packaging of RunAdvisorBot
