+++ src/main/python/advisor/Telegram/Messanger.py
@@ -1,14 +1,22 @@
-import venv
+from dotenv import load_dotenv
+import os, threading, asyncio
 from telegram.ext import Application, CommandHandler, ContextTypes
 from telegram import Update
 import requests
 
 class TelegramMessenger:
     def __init__(self, chat_id=None):
-        self.BOT_TOKEN = "TOKEN"  # 🛠️ Use environment variable or provided token
+        load_dotenv()  # Load environment variables from .env file
+        self.BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")  # 🛠️ Use environment variable or provided token
+        if not self.BOT_TOKEN:
+            raise ValueError("❌ TELEGRAM_BOT_TOKEN not found in .env file")
+
         self.chat_id = chat_id
         self.should_run = True  # 🔁 Flag to control bot execution
-
+        
+    def run_bot_async(self):
+        threading.Thread(target=self.run_bot, daemon=True).start()
+        
     async def start(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
         chat_id = update.effective_chat.id
         self.chat_id = chat_id
@@ -19,16 +27,27 @@ class TelegramMessenger:
         self.should_run = False
         await context.bot.send_message(chat_id=update.effective_chat.id, text="🛑 Advisor stopped by user.")
 
+    async def status(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
+        if self.should_run:
+            await context.bot.send_message(chat_id=update.effective_chat.id, text="✅ Advisor is running.")
+        else:
+            await context.bot.send_message(chat_id=update.effective_chat.id, text="🛑 Advisor is stopped.")
+            
     def run_bot(self):
         """
         Starts the Telegram bot and waits for commands.
         """
+        asyncio.set_event_loop(asyncio.new_event_loop())  # 🧠 create and set event loop in thread
+        loop = asyncio.get_event_loop()
+
         app = Application.builder().token(self.BOT_TOKEN).build()
         app.add_handler(CommandHandler("start", self.start))
         app.add_handler(CommandHandler("stop", self.stop))  # 🆕 Add stop command
+        app.add_handler(CommandHandler("status", self.status))  # 🆕 Add status command
 
         print("🤖 Bot is running. Use /start to begin and /stop to stop the advisor.")
-        app.run_polling()
+        loop.run_until_complete(app.run_polling())  # 🧠 run polling in the new loop
+
 
     def send_message(self, message):
         if not self.chat_id:
